1.0 b6 (09.03.2024) (pre3) finalising GuitarPro based EoF simple time distribution for new songs
1.0 b6 (09.03.2024) (pre3) finalising GuitarPro based EoF simple time distribution for new songs
[x] fix improved import no eof no gp5
[x] improved getfirst note, also added get last
[x] fix move of buttons import selection of param screen
[x] fix new attribute to always give if songs match standardization rule
[x] fix some empty/broken audio/album paths
[x] fix some original album paths & code at path fix (not sure where the code issues is)
[x] why there are no Packing details in db
[x] standardise the get records specialyl since SQL-Lite might return an empty record=1 if tableempty
[x] save setting doesnt seem to work
[x] 3rd extension was not cleaned (psarc from celand folders
[x] fixes on cleanups
[x] further sqlite improve impl strcom,switch,ucase etc.
	[x] weird pack no
	[x] why zero packed songs
	[x] sqlite dlcmnanger 1700 also made theser calls avaiul in sqlite
[x] where is sample (not the case on single processing as anyway sumamry at end)
[x] sqlite other windows (not required as now selects etc standardized only in utilities fuction files)
[x] improved packing no detection at single and harmogenised with mass same function
[x] harmonised generate param list same function, and added general to all (still to do put it sec :))
[x] generate packing stats ignores platform case
[x] workaround sqlite table with id not null contrains not being able to insert in even if not inserting id
[-] add instrumntal AI split https://github.com/stemrollerapp/stemroller
[-] fix search 2-3times
[-] full album not detected
[-] setting attribute no displayed
[-] fix acdc
[-] check full files
[-] check for file size zero
[-] if preview missing check add full song as
[-] add weird char from file name to path an
[-] add remove doublefiled wems
[-] check twhat wems i deleted
[-] check wehat bnk still are after cleanup
[x] Has_Alternate_Lyrics
[x] Has_Alternate_Audio
[x] a440
[x] ukulele
[x] Metallcover
[x] mix
[x]  make attributes dynamically position  themselves
[x] when changig metadata give a change to edit
[x] remaster not detected on album (added)
[-] dont add packing grp to meta
[x] fix last meta
[x] check digitec on torn cover by neck
[x] strech few maindb
[x] increse font in error window
[x] fix pack regress issue
[x] move call to proced as to show progress
[x] fixstandardization screen zoom
[x] fixstandardization screen load
[x] open standard doesnt getsu top the artist&album
[x] pack no doenst increase(failed pack is the same untill increased) (maybe get max and insease then)its ok as isO)
[x] fix add attribute through right click
[x] if indiv repack is not in failed dont mark as broken :)
[x] reinforced packing audio file using logic (mostly cause of ps3 issues) and error reporting out of it
[x] enhance right clicks on the 3 right click menu grps
[x] fix duplicate groups
[x] if you dont add stuff to lyrics display a warning :)(ConfigRepository.Instance()["dlcm_AdditionalManipul73"] == "Yes" && filez.Has_Vocals == "Yes" && c("dlcm_AdditionalManipul102") != "Yes")
[x] increase song details to max-1 in lyrics
[x] add def dlc_id_
[x] dlcname should be part of getmeta
[x] (qnap modif file) fix bass dankoC:\t\0\0_data\Pc_CDLC_Danko Jones_2015_Fire Music_0_Gonna Be A Fight Tonight_12128\songs\arr\fakk_djtgbaft_bass.xml
[x] add title to error window
[x] 	Failed at packing (1010970) :4 (inprov clean path of umlaut chars)
[x] fix sections not being gathered since 2021 :( :( :( :( :(
[x] add regather attributes (tunning frecv, sections++later)
[x] fixed no doubt sunda qnap broken bass
[x] C:\t\0\0_data\Pc_CDLC_Tommy Johansson_2022_I'm Still Standing_0_I'm Still Standing_14244 fixed missin audio and bnk
[x] C:\t\0\0_data\Pc_CDLC_Rise Against_2001_The Unraveling_0_Faint Resemblance_7583\manifests\songs_dlc_cusrafaintresemblance fixed empty json
[x] fix dlcm opening
[-] fix import when w duplic doenst import new
[] why grp is not correct for smnoko
[] whatsw eof

+++ RocksmithToolkitLib/DLCPackage/DLCPackageCreator.cs
@@ -105,6 +105,7 @@ namespace RocksmithToolkitLib.DLCPackage
         /// <param name="pnum">Packages remaining to generate (used to control art cache)</param>
         public static string Generate(string destPath, DLCPackageData info, Platform platform, DLCPackageType dlcType = DLCPackageType.Song, int pnum = -1)
         {
+            var er = "";
             switch (platform.platform)
             {
                 case GamePlatform.XBox360:
@@ -131,7 +132,7 @@ namespace RocksmithToolkitLib.DLCPackage
                         switch (dlcType)
                         {
                             case DLCPackageType.Song:
-                                GenerateRS2014SongPsarc(packPsarcStream, info, platform, pnum);
+                                er = GenerateRS2014SongPsarc(packPsarcStream, info, platform, pnum);
                                 break;
                             case DLCPackageType.Lesson:
                                 throw new NotImplementedException("Lesson package type not implemented yet :(");
@@ -183,7 +184,7 @@ namespace RocksmithToolkitLib.DLCPackage
                         if (!archivePath.EndsWith(".psarc"))
                             archivePath += ".psarc";
 
-                        archivePath = EncryptPS3EdatFiles(archivePath, platform);
+                        archivePath = EncryptPS3EdatFiles(archivePath, platform); //bcapi dont enrcypt if failed at packing
                         break;
                 }
             }
@@ -201,6 +202,7 @@ namespace RocksmithToolkitLib.DLCPackage
             if (pnum <= 1)// doesn't trigger for last one, should be ? == 1 when last package generated.
                 DeleteTmpFiles(TMPFILES_ART);
 
+            if (!File.Exists(archivePath)) archivePath="missing file: "+er;
             return archivePath;
         }
 
@@ -318,27 +320,29 @@ namespace RocksmithToolkitLib.DLCPackage
                     File.Delete(ps3File);
 
             // Move directory if RS1 or file in RS2014 to user selected path
-            if (platform.version == GameVersion.RS2014)
-            {
-                var encryptedFile = String.Format("{0}.edat", FILES_PS3[0]);
-                destPath = String.Format("{0}.edat", srcPath);
+            if (FILES_PS3.Count > 0)
+                if (platform.version == GameVersion.RS2014)
+                {
+                    var encryptedFile = String.Format("{0}.edat", FILES_PS3[0]);
+                    destPath = String.Format("{0}.edat", srcPath);
 
-                if (File.Exists(destPath))
-                    File.Delete(destPath);
+                    if (File.Exists(destPath))
+                        File.Delete(destPath);
 
-                if (File.Exists(encryptedFile))
-                    File.Move(encryptedFile, destPath);
-            }
-            else
-            {
-                if (Directory.Exists(PS3_WORKDIR))
-                    IOExtension.MoveDirectory(PS3_WORKDIR, String.Format("{0}_PS3", srcPath), true);
-            }
+                    if (File.Exists(encryptedFile))
+                        File.Move(encryptedFile, destPath);
+                }
+                else
+                {
+                    if (Directory.Exists(PS3_WORKDIR))
+                        IOExtension.MoveDirectory(PS3_WORKDIR, String.Format("{0}_PS3", srcPath), true);
+                }
 
             if (encryptResult.IndexOf("No JDK or JRE is installed on your machine") > 0)
                 throw new InvalidOperationException("You need install Java SE 7 (x86) or higher on your machine. The Java path should be in PATH Environment Variable:" + Environment.NewLine + Environment.NewLine + encryptResult);
 
-            if (encryptResult.IndexOf(Packer.EDAT_MSG) < 0)
+            if (!File.Exists(destPath)) return destPath;
+            else if (encryptResult.IndexOf(Packer.EDAT_MSG) < 0)
                 throw new InvalidOperationException("Rebuilder error, please check if .edat files are created correctly and see output bellow:" + Environment.NewLine + Environment.NewLine + encryptResult);
 
             return destPath;
@@ -350,9 +354,10 @@ namespace RocksmithToolkitLib.DLCPackage
 
         #region Generate PSARC RS2014
 
-        public static void GenerateRS2014SongPsarc(Stream output, DLCPackageData info, Platform platform, int pnum = -1)
+        public static string GenerateRS2014SongPsarc(Stream output, DLCPackageData info, Platform platform, int pnum = -1)
         {
             // TODO: Benchmark processes and optimize speed
+            var er = "";
             dlcName = info.Name.ToLower();
             packPsarc = new PSARC.PSARC();
 
@@ -662,7 +667,7 @@ namespace RocksmithToolkitLib.DLCPackage
                 //}
                 //string ss = String.Format("Error 2generate PS3 package: {0}{1}. {0}PS3 package require 'JAVA x86' (32 bits) installed on your machine to generate properly.{0}", Environment.NewLine, ex.StackTrace);
                 //MessageBox.Show(ex.Message);
-                var tgst = "Error during rocksmith lib internal packaging..." + ex; UpdateLog(DateTime.Now, tgst, false, ConfigRepository.Instance()["dlcm_TempPath"], null, null, null, null);
+                er = "Erro during rocksmith lib internal packaging..." + ex.Message; UpdateLog(DateTime.Now, er, false, ConfigRepository.Instance()["dlcm_TempPath"], null, null, null, null);
                 //errorsFound.AppendLine(ss);
             }
             finally
@@ -680,6 +685,7 @@ namespace RocksmithToolkitLib.DLCPackage
                     DeleteTmpFiles(TMPFILES_ART);
                 DeleteTmpFiles(TMPFILES_SNG);
             }
+            return er;
         }
 
         public static DateTime UpdateLog(DateTime dt, string txt, bool bbl, string tmpPath, string MultithreadNo, string form, ProgressBar pB_ReadDLCs, RichTextBox rtxt_StatisticsOnReadDLCs)
@@ -699,7 +705,7 @@ namespace RocksmithToolkitLib.DLCPackage
 
             if (form == "MainDB")
                 ismaindb = "maindb";
-            
+
             Random randomp = new Random();// Write the string to a file. packid+
             var packid = 0;
             packid = randomp.Next(0, 100000);
